Perf: Filter dictionary once; partition dictionary
Filter dictionary once:
Words that are in the dictionary that don't start with any letter in the
acronym are filtered so they don't have to be checked every time. This
is a good compromise between filter by acronym first (too slow but more
consistent) and filter by anagram first (slow but increasingly fast). We
filter by acronym ONCE, then anagram first

Partition dictionary:
Every round of expand, instead of filtering dictionary by acronym and
throwing the work away, we split the dictionary in two: those that fit
the letter and those that don't. Those that don't are passed down in the
new dictionary. Note that this ONLY WORKS in the special case of wfaoie
where NO LETTER IS REPEATED. A check must be made in the general case to
ensure that is so, which is still a TODO.